app/requirements.txt
fastapi
uvicorn
python-multipart
passlib
PyJWT
tensorflow
opencv-python
pillow
numpy
